fix jepa tests
